Moved Defragmenter to New Folder Structure
As discussed in https://github.com/google/gopacket/pull/449, the
defragmenter was moved to a seperate package that will only contain the
LCM defragmenter. This is done in order to enable a more clean package
setup.
