+++ defrag/lcmdefrag/lcmdefrag.go
@@ -4,7 +4,8 @@
 // that can be found in the LICENSE file in the root of the source
 // tree.
 
-package defrag
+// Package lcmdefrag contains a defragmenter for LCM messages.
+package lcmdefrag
 
 import (
 	"fmt"
